#  Accelerate Video Generation, Reducing the Time to Generate a 62s Video from 1000s to About 25s