Main Optimization Points:
- Avoid frequent creation and release of the renderer.
- Reduce the number of concurrent processes (to lessen the overhead of process creation and scheduling), and have each process only send the data it needs to process (to reduce data transfer).
- Use BMP as the temporary image storage format and parallelize the rendering and file-saving processes.
- Employ ffmpeg for media processing, which improves generation speed.
- Adjust some rendering configuration parameters, such as video resolution and FPS, allowing for trade-offs between generation speed and quality according to needs.

Notes:
- For media processing with ffmpeg, please install it first: **conda install -c conda-forge ffmpeg.**
- By default, BMP is used as the temporary image format. Generating a video of 1920x720 30fps for 62s requires 9G of disk space temporarily, and a 1920x1080 30fps 62s video needs 11G of disk space. These temporary images will be deleted after video generation, which I think is not too big of an issue. If the space requirement is considered too high, you can set args.render_tmp_img_filetype to "png" to generate temporary PNG files instead, though this may increase the time required.
- I have named the final output video as res_2_scott_0_1_1.mp4, whereas the original code used final_clip as 2_scott_0_1_1.mp4, which I believe should not cause any issue.
- **I only have Windows, and it has been tested and passed on Windows. It has not been tested on other platforms.**